修复bug和修改readme
