+++ linpg/api/__init__.py
@@ -1,6 +1,6 @@
 # cython: language_level=3
 """
 结构:
-basic -> controller -> module -> font -> tool
+basic -> controller -> module -> system -> font -> tool
 """
 from .tool import *
